// nodejs-react(hotel service)/HotelFrontEnd/frontend/src/styles/style.type.ts
export type StyleProps = {
  display?: string;
  visibility?: string;
  position?: string;
  flex?: string;
  flexWrap?: string;

  overflow?: string;

  top?: string;
  right?: string;
  left?: string;
  bottom?: string;

  flexDirection?: string;
  alignItems?: string;
  justifyContent?: string;
  alignSelf?: string;

  padding?: string;
  margin?: string;

  height?: string;
  width?: string;
  maxWidth?: string;
  minWidth?: string;
  maxHeight?: string;
  minHeight?: string;

  objectFit?: string;

  backgroundColor?: string;
  boxShadow?: string;

  fontFamily?: string;
  fontSize?: string;
  lineHeight?: string;
  fontWeight?: string;
  textAlign?: string;
  letterSpacing?: string;

  color?: string;
  outline?: string;

  border?: string;
  borderRadius?: string;

  placeContent?: string;

  textTransform?: string;

  resize?: string;

  name?: string; //for debug
  zIndex?: string;
};